Move CalculateMoneyDonatedTest to projects app.
The code to update the project is now in the project app so it makes sense to
have the test there as well.

BB-475
